+++ Options/BotOptions.cs
@@ -2,6 +2,6 @@ namespace ReviewGrabberBot.Options
 {
     internal sealed class BotOptions
     {
-        public int AdminId;
+        public long AdminId;
     }
 }
